fix: User.schema
