+++ src/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { PassportModule } from '@nestjs/passport';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { JwtStrategy } from './strategy/jwt.strategy';
-import { User, UserSchema } from '../schemas/user.schema';
+import { User, UserSchema } from '../schemas/User.schema';
 import { UsersRepository } from './users.repository';
 import { HttpModule } from '@nestjs/axios';
 import { KakaoStrategy } from './strategy/kakao.strategy';
