+++ src/schemas/User.schema.ts
@@ -1,20 +1,20 @@
 import { Prop, Schema, SchemaFactory, SchemaOptions } from '@nestjs/mongoose';
 import { Transform } from 'class-transformer';
-import { Document, ObjectId } from 'mongoose';
+import { Document, ObjectId, Types } from 'mongoose';
 
 export type UserDocument = User & Document;
+
 const options: SchemaOptions = {
-  toJSON: {
-    getters: true,
-    virtuals: true,
-  },
+  collection: 'users',
+  timestamps: true,
 };
 @Schema(options)
-export class User {
+export class User extends Document {
   @Transform(({ value }) => value.toString())
   _id: ObjectId;
 
   @Prop({
+    default: null,
     index: { unique: true, dropDups: true },
     type: String,
   })
@@ -41,13 +41,9 @@ export class User {
   position: string;
 
   @Prop({
-    default: null,
+    default: [],
   })
-  createdAt: Date;
-  @Prop({
-    default: null,
-  })
-  updatedAt: Date;
+  dmRooms: [];
 }
 
 export const UserSchema = SchemaFactory.createForClass(User);
